% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported.R
\name{boom}
\alias{boom}
\alias{rig}
\alias{rigger}
\alias{rig_in_place}
\alias{rig_on_load}
\title{Print the Output of Intermediate Steps of a Call}
\usage{
boom(expr, clock = NULL, print = NULL)

rig(fun, clock = NULL, print = NULL)

rigger(clock = NULL, print = NULL)

rig_in_place(..., clock = NULL, print = NULL)

rig_on_load()
}
\arguments{
\item{expr}{call to explode}

\item{clock}{whether to time intermediate steps. Defaults to \code{getOption("boomer.clock")}
evaluated at run time (\code{FALSE} unless you change it). The execution time of
a step doesn't include the execution time of its previously printed sub-steps.}

\item{print}{A function, a formula or a list of functions or formulas, used to
modify the way the output is printed. Defaults to \code{getOption("boomer.print")}
evaluated at run time (\code{base::print} unless you change it)'.}

\item{fun}{function ro \code{rig()}, can be a symbol, an expression returning a function, or a string}

\item{...}{Functions to rig in their namespace

If the \code{print} argument is a function, it will be used to print, or to transform the output
before it's printed. Use \code{invisible} to display nothing, useful possibilities are
\code{constructive::construct}, \code{str} or \code{dplyr::glimpse}.

\emph{{rlang}}'s formula notation is supported, so for instance you can type:
\code{print = ~ dplyr::glimpse(., width = 50)}.

Sometimes you might want to print a specific type of object in a custom way,
in this case you can provide a named list, if you provide an unnamed element
it will be used as the default, and named elements will define how objects
of the given S3 class are printed. For instance \code{print = list(str, data.frame = tibble::as_tibble)}}
}
\value{
\code{boom()} returns the output of the call. \code{rig()} returns the modified
input function. \code{rig_in_namespace()} returns \code{invisible(NULL)} and is called
for side effects. \code{rigger()} returns a list containing the arguments, with
the class "rigger" to enable \code{+.rigger} and \code{print.rigger}
}
\description{
\itemize{
\item \code{boom()} prints the intermediate results of a call or a code chunk.
\item \code{rig()} creates a copy of a function which will display the intermediate
results of all the calls of it body.
\item \code{rig_in_place()} rigs a function in place, so its always
verbose even when called by other existing functions. It works on functions from packages
as well as on functions defined in a session.
To undo, call \code{load_all()} for the development package or
\code{pkgload::unload()} on other packages, or restart the session if your rigged a base package. Shouldn't be used on S3 generics, but works
on S3 methods.
\item \code{rig_on_load()} can be used in \code{.onLoad()} to rig functions (values or names) stored in \code{getOption("boomer.rig_on_load")}
\item \code{rigger()} provides a convenient way to rig an
anonymous function by using the \code{rigger(...) + function(...) {...}} syntax.
}
}
\section{Side effects of \code{rig_in_place()}}{


When called on a packaged function \code{rig_in_place()} replaces the target function in the namespace but
to behave as expected it does a bit more:
\itemize{
\item Replace the copy of the function in the package environment, where the functions called without \code{::} or \code{:::} are fetched from after calling \code{library()}.
\item Replace the copy of the function in the S3 methods table if relevant. S3 dispatch calls copies of functions, not the function in the namespace directly, so those need to be replaced as well.
\item Replace the copy of the function in the "imports" environments of already loaded packages. Indeed packages that import function load copies of those in an environment, these need to be replaced as well.
}
}

\examples{
# explode a simple call
boom(subset(head(mtcars, 2), qsec > 17))

# clock calls and customize how to print output
boom(subset(head(mtcars, 2), qsec > 17), clock = TRUE, print = str)

# print str only for data frames
boom(subset(head(mtcars, 2), qsec > 17), print = list(data.frame = str))

# rig an existing function
rig(ave)(warpbreaks$breaks, warpbreaks$wool)

# rig an anonymous function
fun1 <- rigger() + function(x) x + 1 + 2 # same as rig(function(x) x + 1 + 2))
fun1(1)
fun2 <- rigger(TRUE, typeof) + function(x) x + 1 + 2
fun2(1)
}
