% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bquote2.R
\name{bquote2}
\alias{bquote2}
\title{Partial substitution in expressions}
\usage{
bquote2(expr, where = parent.frame(), splice = FALSE)
}
\arguments{
\item{expr}{A \link[base]{language object}.}

\item{where}{An environment.}

\item{splice}{Logical; if \code{TRUE} splicing is enabled.}
}
\description{
Modified \code{bquote()} to interpolate code conditionally in a readable way.
Use \code{.IF(cond, yes, no)} for conditional interpolation, \code{cond} will be evaluated
and \code{yes} and \code{no} will be quoted. \code{no} is optional.
}
\examples{
fun <- function(x, y, cond) {
  boomer:::bquote2({
    some_long_code_chunk_1(.(x))
    .IF(cond, do_something())
    some_long_code_chunk_2(.(y))
  })
}
fun(1, 2, TRUE)
fun(1, 2, FALSE)
}
\keyword{internal}
