% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported.R
\name{rig_in_namespace}
\alias{rig_in_namespace}
\title{rig_in_namespace}
\usage{
rig_in_namespace(..., clock = NULL, print = NULL)
}
\description{
\code{rig_in_namespace()} is a deprecated alias to \code{rig_in_place()}
}
\keyword{internal}
