% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{rig_shiny}
\alias{rig_shiny}
\title{Debug a shiny app using boomer}
\usage{
rig_shiny(
  funs = c("reactive", "renderCachedPlot", "renderDataTable", "renderImage",
    "renderPlot", "renderPrint", "renderTable", "renderText", "renderUI"),
  select = TRUE
)
}
\arguments{
\item{funs}{The reactive functions to be affected, by default \code{reactive()} and
all \verb{render*()} functions from the \emph{{shiny}} package.}

\item{select}{Whether to debug all reacives (\code{TRUE} by default)}
}
\description{
\code{rig_shiny()} should be called after your \code{library(shiny)} call,
it will change your global environment so you might want to refresh your session
after using it. Note that \code{renderPrint()} captures the printed output and in
doing so diverts boomer's ouput from the console to the app. We might find a
way to solve it at some point.
}
