% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported.R
\docType{data}
\name{rigger}
\alias{rigger}
\title{Create rigged function conveniently}
\format{
An object of class \code{rigger} of length 0.
}
\usage{
rigger
}
\description{
Allows \code{rigger + function(...) {...}} syntax to create a rigged function
conveniently
}
\keyword{datasets}
