% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported.R
\name{boom}
\alias{boom}
\alias{rig}
\title{Print the Output of Intermediate Steps of a Call}
\usage{
boom(
  expr,
  clock = getOption("boom.clock"),
  print = getOption("boom.print"),
  ignore = getOption("boom.ignore")
)

rig(
  fun,
  clock = getOption("boom.clock"),
  print = getOption("boom.print"),
  ignore = getOption("boom.ignore")
)
}
\arguments{
\item{expr}{call to explode}

\item{clock}{whether to time intermediate steps, \code{FALSE} by default unless you
set \code{options(boom.clock = TRUE)}. The execution time of a step doesn't include the
execution time of its previously printed sub-steps.}

\item{print}{A function, a formula or a list of functions or formulas.}

\item{fun}{function ro \code{rig()}}
}
\description{
\code{boom()} will explode a call into the output of its parts.
\code{rig()} a function to make all the calls of its body \code{boom()}.
}
\details{
By default, unless the "boom.print" option  is set to a custom value, the
output of every step is printed in a standard way.

If you provide another function such, for instance \code{options(boom.print = str)}
the console output of \code{str} will be printed. Use \code{invisible} to display
nothing, another useful alternative would be \code{dplyr::glimpse}.

\emph{{rlang}}'s formula notation is supported, so for instance you can type:
\code{print = ~ dplyr::glimpse(., width = 50)}.

Sometimes you might want to print a specific type of object in a custom way,
in this case you can provide a named list, if you provide an unnamed element
it will be used as the default, and named elements will define how objects
of the given S3 class are printed. For instance \code{print = list(str, data.frame = tibble::as_tibble)}

Modifying the options is especially useful if you want to use the addin
with custom behavior.
}
\examples{
# explode a simple call
boom(subset(head(mtcars, 2), qsec > 17))

# clock calls and customize how to print output
boom(subset(head(mtcars, 2), qsec > 17), clock = TRUE, print = str)

# print str only for data frames
boom(subset(head(mtcars, 2), qsec > 17), print = list(data.frame = str))

# rig a function
rig(ave)(warpbreaks$breaks, warpbreaks$wool)
}
