% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported.R
\name{rig_in_namespace}
\alias{rig_in_namespace}
\title{rig functions directly in their namespace}
\usage{
rig_in_namespace(
  ...,
  clock = getOption("boom.clock"),
  print = getOption("boom.print"),
  ignore = getOption("boom.ignore"),
  visible_only = getOption("boom.visible_only")
)
}
\arguments{
\item{...}{functions to rig}

\item{clock}{whether to time intermediate steps, \code{FALSE} by default unless you
set \code{options(boom.clock = TRUE)}. The execution time of a step doesn't include the
execution time of its previously printed sub-steps.}

\item{print}{A function, a formula or a list of functions or formulas.}

\item{ignore}{functions to ignore, defaults to \code{c("~", "\\{", "(", "<-", "<<-", "=")}
unless the option \code{"boom.ignore"} is modified. \code{::} and \code{:::} are always ignored.}

\item{visible_only}{whether functions returning invisibly should be considered,
by default they are unless the option \code{"boom.visible_only"} is set to \code{TRUE}.}
}
\description{
To recover original functions, restart the R session
}
