% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported.R
\name{rig_in_namespace}
\alias{rig_in_namespace}
\title{rig functions directly in their namespace}
\usage{
rig_in_namespace(..., clock = NULL, print = NULL)
}
\arguments{
\item{...}{functions to rig}

\item{clock}{whether to time intermediate steps. Defaults to \code{getOption("boomer.clock")}
evaluated at run time (\code{FALSE} unless you change it). The execution time of
a step doesn't include the execution time of its previously printed sub-steps.}

\item{print}{A function, a formula or a list of functions or formulas, used to
modify the way the output is printed. Defaults to \code{getOption("boomer.print")}
evaluated at run time (\code{base::print} unless you change it)'.

If the \code{print} argument is a function, it will be used to print, or to transform the output
before it's printed. Use \code{invisible} to display nothing, useful possibilities are
\code{str} or \code{dplyr::glimpse}.

\emph{{rlang}}'s formula notation is supported, so for instance you can type:
\code{print = ~ dplyr::glimpse(., width = 50)}.

Sometimes you might want to print a specific type of object in a custom way,
in this case you can provide a named list, if you provide an unnamed element
it will be used as the default, and named elements will define how objects
of the given S3 class are printed. For instance \code{print = list(str, data.frame = tibble::as_tibble)}}
}
\description{
To recover original functions, restart the R session
}
