% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{rig_shiny}
\alias{rig_shiny}
\alias{unrig_shiny}
\title{Debug a shiny app using boomer}
\usage{
rig_shiny()

unrig_shiny()
}
\description{
\code{rig_shiny()} modifies the way shiny works so that a debugging tab will be
added to the shiny app and selected reactive functions will log their
refreshed boomed ouput to the console. \code{unrig_shiny()} undoes his action.
}
