% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported.R
\name{rigger}
\alias{rigger}
\title{Create rigged function conveniently}
\usage{
rigger(clock = getOption("boom.clock"), print = getOption("boom.print"))
}
\arguments{
\item{clock}{whether to time intermediate steps, \code{FALSE} by default unless you
set \code{options(boom.clock = TRUE)}. The execution time of a step doesn't include the
execution time of its previously printed sub-steps.}

\item{print}{A function, a formula or a list of functions or formulas.}
}
\value{
a list containing the arguments, with the class "rigger" to enable
\code{+.rigger} and \code{print.rigger}
}
\description{
Allows \code{rigger(...) + function(...) {...}} syntax to create a rigged function
conveniently.
}
\examples{
fun1 <- rigger() + function(x) x + 1 + 2
fun1(1)
fun2 <- rigger(TRUE, typeof) + function(x) x + 1 + 2
fun2(1)
}
