% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported.R
\name{rigger}
\alias{rigger}
\title{Create rigged function conveniently}
\usage{
rigger(
  clock = getOption("boom.clock"),
  print = getOption("boom.print"),
  ignore = getOption("boom.ignore"),
  visible_only = getOption("boom.visible_only"),
  print_args = getOption("boom.print_args")
)
}
\arguments{
\item{clock}{whether to time intermediate steps, \code{FALSE} by default unless you
set \code{options(boom.clock = TRUE)}. The execution time of a step doesn't include the
execution time of its previously printed sub-steps.}

\item{print}{A function, a formula or a list of functions or formulas.}

\item{ignore}{functions to ignore, defaults to \code{c("~", "\\{", "(", "<-", "<<-", "=")}
unless the option \code{"boom.ignore"} is modified. \code{::} and \code{:::} are always ignored.}

\item{visible_only}{whether functions returning invisibly should be considered,
by default they are unless the option \code{"boom.visible_only"} is set to \code{TRUE}.}

\item{print_args}{whether to print the values of the arguments provided to a rigged
function. They are printed when they are evaluated, if they are evaluated.}
}
\value{
a list containing the arguments, with the class "rigger" to enable
\code{+.rigger} and \code{print.rigger}
}
\description{
Allows \code{rigger(...) + function(...) {...}} syntax to create a rigged function
conveniently.
}
\examples{
fun1 <- rigger() + function(x) x + 1 + 2
fun1(1)
fun2 <- rigger(TRUE, typeof) + function(x) x + 1 + 2
fun2(1)
}
